Move attachments out of common and into its own features section. Purely file moves
