+++ app/lib/common/dialogs/attachment_confirmation.dart
@@ -1,6 +1,6 @@
 import 'package:acter/common/models/types.dart';
 
-import 'package:acter/common/widgets/attachments/post_attachment_selection.dart';
+import 'package:acter/features/attachments/widgets/post_attachment_selection.dart';
 import 'package:acter_flutter_sdk/acter_flutter_sdk_ffi.dart'
     show AttachmentsManager;
 import 'package:flutter/material.dart';
