+++ app/lib/common/dialogs/attachment_selection.dart
@@ -3,7 +3,7 @@ import 'dart:io';
 import 'package:acter/common/dialogs/attachment_confirmation.dart';
 import 'package:acter/common/models/types.dart';
 import 'package:acter/common/utils/utils.dart';
-import 'package:acter/common/widgets/attachments/attachment_options.dart';
+import 'package:acter/features/attachments/widgets/attachment_options.dart';
 import 'package:acter_flutter_sdk/acter_flutter_sdk_ffi.dart'
     show AttachmentsManager;
 import 'package:file_picker/file_picker.dart';
