+++ app/lib/features/attachments/widgets/post_attachment_selection.dart
@@ -1,7 +1,7 @@
 import 'dart:typed_data';
 
 import 'package:acter/common/models/types.dart';
-import 'package:acter/common/widgets/attachments/attachment_draft_item.dart';
+import 'package:acter/features/attachments/widgets/attachment_draft_item.dart';
 import 'package:acter/features/home/providers/client_providers.dart';
 import 'package:acter_flutter_sdk/acter_flutter_sdk_ffi.dart'
     show AttachmentDraft, AttachmentsManager;
