+++ app/lib/features/attachments/widgets/views/video_view.dart
@@ -1,20 +1,17 @@
 import 'package:acter/common/models/attachment_media_state/attachment_media_state.dart';
-import 'package:acter/common/providers/attachment_providers.dart';
-import 'package:acter/common/widgets/image_dialog.dart';
+import 'package:acter/features/attachments/providers/attachment_providers.dart';
+import 'package:acter/common/widgets/acter_video_player.dart';
+import 'package:acter/common/widgets/video_dialog.dart';
 import 'package:acter_flutter_sdk/acter_flutter_sdk_ffi.dart' show Attachment;
 import 'package:flutter/material.dart';
 import 'package:flutter_chat_ui/flutter_chat_ui.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-// Image Attachment preview
-class ImageView extends ConsumerWidget {
+// Video attachment preview
+class VideoView extends ConsumerWidget {
   final Attachment attachment;
   final bool? openView;
-  const ImageView({
-    super.key,
-    required this.attachment,
-    this.openView = true,
-  });
+  const VideoView({super.key, required this.attachment, this.openView = true});
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
@@ -22,9 +19,9 @@ class ImageView extends ConsumerWidget {
     if (mediaState.mediaLoadingState.isLoading || mediaState.isDownloading) {
       return loadingIndication(context);
     } else if (mediaState.mediaFile == null) {
-      return imagePlaceholder(context, attachment, mediaState, ref);
+      return videoPlaceholder(context, attachment, mediaState, ref);
     } else {
-      return imageUI(context, mediaState);
+      return videoUI(context, mediaState);
     }
   }
 
@@ -36,7 +33,7 @@ class ImageView extends ConsumerWidget {
     );
   }
 
-  Widget imagePlaceholder(
+  Widget videoPlaceholder(
     BuildContext context,
     Attachment attachment,
     AttachmentMediaState mediaState,
@@ -50,9 +47,9 @@ class ImageView extends ConsumerWidget {
             context: context,
             barrierDismissible: false,
             useRootNavigator: false,
-            builder: (ctx) => ImageDialog(
+            builder: (ctx) => VideoDialog(
               title: msgContent.body(),
-              imageFile: mediaState.mediaFile!,
+              videoFile: mediaState.mediaFile!,
             ),
           );
         } else {
@@ -79,7 +76,7 @@ class ImageView extends ConsumerWidget {
               mainAxisSize: MainAxisSize.min,
               children: [
                 const Icon(
-                  Icons.image,
+                  Icons.video_file,
                   size: 14,
                 ),
                 const SizedBox(width: 5),
@@ -96,50 +93,25 @@ class ImageView extends ConsumerWidget {
     );
   }
 
-  Widget imageUI(BuildContext context, AttachmentMediaState mediaState) {
+  Widget videoUI(BuildContext context, AttachmentMediaState mediaState) {
+    final msgContent = attachment.msgContent();
     return InkWell(
       onTap: openView!
-          ? () {
-              final msgContent = attachment.msgContent();
-              showAdaptiveDialog(
+          ? () => showAdaptiveDialog(
                 context: context,
                 barrierDismissible: false,
                 useRootNavigator: false,
-                builder: (ctx) => ImageDialog(
+                builder: (ctx) => VideoDialog(
                   title: msgContent.body(),
-                  imageFile: mediaState.mediaFile!,
+                  videoFile: mediaState.mediaFile!,
                 ),
-              );
-            }
+              )
           : null,
       child: ClipRRect(
         borderRadius: BorderRadius.circular(6),
-        child: Image.file(
-          mediaState.mediaFile!,
-          frameBuilder: (
-            BuildContext context,
-            Widget child,
-            int? frame,
-            bool wasSynchronouslyLoaded,
-          ) {
-            if (wasSynchronouslyLoaded) {
-              return child;
-            }
-            return AnimatedOpacity(
-              opacity: frame == null ? 0 : 1,
-              duration: const Duration(seconds: 1),
-              curve: Curves.easeOut,
-              child: child,
-            );
-          },
-          errorBuilder: (
-            BuildContext context,
-            Object url,
-            StackTrace? error,
-          ) {
-            return Text('Could not load image due to $error');
-          },
-          fit: BoxFit.cover,
+        child: ActerVideoPlayer(
+          hasPlayerControls: false,
+          videoFile: mediaState.mediaFile!,
         ),
       ),
     );
