+++ app/lib/features/chat/widgets/custom_input.dart
@@ -4,8 +4,8 @@ import 'package:acter/common/models/types.dart';
 import 'package:acter/common/providers/room_providers.dart';
 import 'package:acter/common/themes/app_theme.dart';
 import 'package:acter/common/utils/utils.dart';
-import 'package:acter/common/widgets/attachments/attachment_container.dart';
-import 'package:acter/common/widgets/attachments/attachment_options.dart';
+import 'package:acter/features/attachments/widgets/attachment_container.dart';
+import 'package:acter/features/attachments/widgets/attachment_options.dart';
 import 'package:acter/common/widgets/emoji_picker_widget.dart';
 import 'package:acter/common/widgets/frost_effect.dart';
 import 'package:acter/features/chat/models/chat_input_state/chat_input_state.dart';
